requirements.txt
grpcio==1.78.0
grpcio-tools==1.78.0
mysql-connector-python==9.6.0
Pillow==12.1.1
protobuf==6.33.5
pystray==0.19.5
python-dotenv==1.2.2
setuptools==82.0.1
six==1.17.0
typing_extensions==4.15.0

# macOS only
pyobjc-core==12.1; sys_platform == "darwin"
pyobjc-framework-Cocoa==12.1; sys_platform == "darwin"
pyobjc-framework-Quartz==12.1; sys_platform == "darwin"

# Windows only
WMI==1.4.9; sys_platform == "win32"
pywin32==310; sys_platform == "win32"